Adjust CalculateRemainingBudgetResponse return value, created CalculateRemainingBudget class and adjusted Budget so ReturnTotalRemainingBudget returns a double

+++ src/main/java/entity/Budget.java
@@ -47,8 +47,8 @@ public class Budget {
         return this.totalBudget;
     }
 
-    public int getTotalRemainingBudget() {
-        int sum = 0;
+    public double getTotalRemainingBudget() {
+        double sum = 0;
         for (Double value : this.remainingBudget.values()) {
             sum += value;
         }
